backend/requirements-simple.txt
# IPv6 WireGuard Manager - 简化依赖
# 核心依赖
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# 数据库
sqlalchemy==2.0.23
alembic==1.13.1
pymysql==1.1.0
aiomysql==0.2.0

# 认证和安全
python-jose[cryptography]>=3.3.0
passlib[argon2,bcrypt]>=1.7.4
argon2-cffi>=21.0.0
python-multipart>=0.0.6

# HTTP客户端
httpx==0.25.2

# 工具库
python-dotenv==1.0.0
click==8.1.7

# WireGuard相关
cryptography>=41.0.0,<47.0.0
qrcode[pil]==7.4.2

# 系统监控
psutil==5.9.6

# 邮件验证
email-validator==2.1.0
